// src/app/model/outcome-model.ts
export class outcomeModel {
    id: number=0;
    fecha: Date = new Date();
    tipoPago: string='';
    factura: string='';
    estado: boolean=false;
    clientId: number=0;
    clientRazonSocial: string='';
    clientRuc: number=0;
    clientCorreo: string='';
    clientContacto: number=0;
    clientDireccion: string='';
    userId: number=0;
    userNombre: string='';
    userAuthId: number=0;
	userAuthNombre: string='';
}