// src/app/model/productos_model.ts
  
  export class ProductModel{
    id: number=0;
    nombre: string = '';
    color: string = '';
    precio: number=0.0;
    stock: number=0;
    estado: boolean = false;
    categoryId: number=0;
    categoryNombre: string = '';
    supplierId: number=0;
    supplierRazonSocial: string = '';
    supplierRuc: number=0;
    supplierContacto: number=0;
  }